+++ src/Samples/src/samples/particles/triangles_sps_scene.cpp
@@ -77,7 +77,7 @@ public:
     _sps                 = SolidParticleSystem::New("SPS", scene, spsOptions);
     SolidParticleSystemMeshBuilderOptions spsMeshBuilderOptions;
     spsMeshBuilderOptions.positionFunction
-      = [this](SolidParticle* particle, unsigned int i, unsigned int s) -> void {
+      = [this](SolidParticle* particle, size_t i, size_t s) -> void {
       _positionFunction(particle, i, s);
     };
     _sps->addShape(triangle, _nb, spsMeshBuilderOptions);
@@ -98,7 +98,7 @@ private:
   /**
    * @brief Custom position function for SPS creation.
    */
-  void _positionFunction(SolidParticle* particle, unsigned int /*i*/, unsigned int /*s*/)
+  void _positionFunction(SolidParticle* particle, size_t /*i*/, size_t /*s*/)
   {
     particle->position.x = (Math::random() - 0.5f) * _fact;
     particle->position.y = (Math::random() - 0.5f) * _fact;
